Move initializer test to the proper file
