+++ tests/keras/initializers_test.py
@@ -104,6 +104,18 @@ def test_orthogonal(tensor_shape):
             target_mean=0.)
 
 
+def test_orthogonal_init_does_not_affect_global_rng():
+    np.random.seed(1337)
+    before = np.random.randint(0, 100, size=10)
+
+    np.random.seed(1337)
+    init = initializers.orthogonal(seed=9876)
+    init(shape=(10, 5))
+    after = np.random.randint(0, 100, size=10)
+
+    assert np.array_equal(before, after)
+
+
 @pytest.mark.parametrize('tensor_shape',
                          [(100, 100), (10, 20), (30, 80), (1, 2, 3, 4)],
                          ids=['FC', 'RNN', 'RNN_INVALID', 'CONV'])
